pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "avacore"
version = "0.8.2"
description = "Local personal assistant core for Ava"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "pydantic>=2.8",
  "python-dotenv>=1.0",
  "requests>=2.32",
  "aiosqlite>=0.20",
  "python-telegram-bot>=21.0",
  "pymupdf>=1.24",
  "sentence-transformers>=3.0",
  "numpy>=1.26,<2",
  "faiss-cpu>=1.8",
  "feedparser>=6.0",
  "transformers>=4.50.0.dev0,<4.58",
  "torch>=2.2",
  "accelerate>=1.0",
  "pillow>=10.0",
  "num2words>=0.5.14"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "mypy>=1.11"
]

[tool.setuptools.packages.find]
include = ["avacore*"]
exclude = ["data*", "tests*", "scripts*"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
